backend/requirements.txt
# SentimentAI Backend Dependencies
# Production-grade requirements with pinned versions

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9  # PostgreSQL adapter for Supabase
alembic>=1.13.0  # Database migrations

# Machine Learning
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.26.0
joblib>=1.3.0

# Logging & Monitoring
loguru>=0.7.2
prometheus-client>=0.19.0
prometheus-fastapi-instrumentator>=6.1.0
python-json-logger>=2.0.7

# Utilities
python-multipart>=0.0.6  # File uploads
python-dotenv>=1.0.0  # Environment variables

# Phase 1: Security & Infrastructure
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.9
bleach>=6.1.0

# Phase 2: MLOps
mlflow>=2.9.2
evidently>=0.4.11
redis>=5.0.1
apscheduler>=3.10.4

# Phase 5: Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
locust>=2.19.1

# Development
black>=23.12.0
flake8>=6.1.0
pre-commit>=3.6.0
